+++ CHANGELOG.md
@@ -17,6 +17,8 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 - check-ssl-qualys.rb: Handle nil grade_rank as critical not rated
 
 
+- check-ssl-crl.rb: Test for expiring CRL
+
 ## [1.0.0]
 ### Changed
 - Updated Rubocop to 0.40, applied auto-correct